Move workspace tree and thumbnails components
