+++ src/app/pages/workspace/components/json-tree/json-tree.component.scss
