+++ src/app/pages/workspace/components/page-thumbnails/page-thumbnails.component.scss
